imp: kill: target children of a fellow flux-imp
Problem: If a flux-imp lingers as parent of a job shell, it would
not do to send the IMP process signals directed at the job and thus
the job shell. However, the PID of the flux-imp will be the target
passed by the host instance to `flux-imp kill`.

Capture the command name in the `struct pid_info` so that the IMP
can determine if a root owned flux-imp process is the target of the
kill command. If this is case, target the fellow IMP's children
instead.