configure: add --enable-pam option to link IMP with libpam
Problem: The flux-imp executable does not link with libpam, but this
will be required in order to add PAM support to IMP exec.

Add an X_AC_PAM configure macro and use it to check for libpam
and libpam_misc if --enable-pam is passed to configure.

Add PAM_LIBS to flux-imp build so that the IMP is linked with PAM
when --enable-pam is used.