testsuite: add t2003-imp-exec-pam.t
Problem: There are no tests in the testsuite that verify functionality
of a PAM-enabled IMP.

Add t2003-imp-exec-pam.t which contains this testing. All tests are
skipped if the IMP was not built with PAM support or libpam_wrapper.so
is not found in the default ld search path.

The test uses libpam_wrapper.so to allow testing with a generated
PAM config, instead of requiring modification of system configuration.