imp: use helper for IMP input when FLUX_IMP_EXEC_HELPER is set
Problem: The IMP is only capable of reading its "input" on stdin,
which means the Flux execution system must write this data to each
IMP process invoked for a job. For large jobs, this could become a
scalability problem. Additionally, there are certain use cases where
it would be useful to keep the stdio descriptors free in the job shell,
but this method for reading input makes that impossible.

Support an optional FLUX_IMP_EXEC_HELPER environment variable in the
IMP which, when set, causes the IMP to invoke the specified helper
program to gather the input, instead of expecting input on stdin.

The value in FLUX_IMP_EXEC_HELPER is passed directly to safe_popen(),
so it will be tokenized as a command line separated by whitespace,
but no other substitutions or processing is performed as would happen
with the shell for instance. This reduces the probability of bugs
introduced by the FLUX_IMP_EXEC_HELPER environment variable, while
still allowing some flexibility in the command line, for instance

 FLUX_IMP_EXEC_HELPER="flux imp-exec-helper"

will work as expected.